requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==4.2
django-celery==3.1.17
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
kombu==5.3.7
Markdown==3.6
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
